+++ src/components/Overlays/Modal/index.ts
@@ -1,1 +1,2 @@
 export { Modal } from './Modal'
+export type { IModal, IModalConfig, IModalOpenRequestData } from './ModalTypes'