+++ src/components/SelectBox/index.ts
@@ -1,2 +1,3 @@
 export { SelectBox } from './SelectBox'
+export type { SelectBoxProps } from './SelectBox'
 export { useSelectBox } from './hooks/useSelectBox'