+++ src/components/Table/ColGroup.tsx
@@ -1,0 +1,7 @@
+export interface ColGroupProps extends React.ComponentPropsWithoutRef<'colgroup'> {
+  children: React.ReactNode
+}
+
+export function ColGroup({ children, ...props }: ColGroupProps) {
+  return <colgroup {...props}>{children}</colgroup>
+}